refactor: Update project and client interfaces for consistency and remove unused types
